requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlmodel>=0.0.14
python-multipart>=0.0.6
httpx>=0.25.1
pillow>=10.4.0
python-dateutil>=2.8.2
pydantic>=2.5.0
python-dotenv>=1.0.0
